refactor(configurator): split into BundleStateManager + BundleInspector + facade
Extract the monolithic SessionConfigurator (1 898 lines) into three focused
modules:

- configurator/_provenance_utils.py: module-level helpers (_PROV_CATEGORY_MAP,
  _normalize_module_name, _build_normalized_prov_lookup, _lookup_prov_behavior)
  shared by both the state manager and the inspector.

- configurator/_state_manager.py: BundleStateManager — all mutation methods
  (context/agent/tool/provider/hook/behavior enable/disable, config_set,
  save, apply_saved_settings) plus the hook-capture, module-to-tool, and
  hook-handler-to-module mappings built at init.  Exposes read-only
  properties for the inspector.

- configurator/_inspector.py: BundleInspector — all query methods
  (snapshot/diff_from_original/take_snapshot and all *_list() methods).
  Takes a BundleStateManager reference; owns _original_snapshot.

- configurator/__init__.py: ~170 line facade that creates the two objects,
  delegates every public method as a one-liner, re-exports the helper
  functions so all existing imports continue to work, and exposes backward-
  compat property accessors for private attributes accessed by tests.
